auth reastriction

+++ Controllers/CommentController.php
@@ -2,10 +2,11 @@
 
 namespace App\Controllers;
 
+use App\Utils\Alert;
+use App\Models\PostManager;
+use App\Models\UserManager;
 use App\Models\CommentManager;
 use App\Controllers\CoreController;
-use App\Models\PostManager;
-use App\Utils\Alert;
 
 final class CommentController extends CoreController
 {
@@ -16,57 +17,83 @@ final class CommentController extends CoreController
 	{
 		$this->commentManager = new CommentManager();
 		$this->postManager = new PostManager();
+		$this->userManager = new UserManager();	
 	}
 
 	public function insert()
 	{
-		$this->commentManager->insert((string)$_POST['content'], (int)$_POST['post_id']);
-		Alert::addAlert(Alert::GREEN, 'Enregistrement du commentaire effectué.');
-		$this->redirect("show-post", ["id" => $_POST['post_id']]);
+		if($this->isLogged()){
+			$this->commentManager->insert((string)$_POST['content'], (int)$_POST['post_id']);
+			Alert::addAlert(Alert::GREEN, 'Enregistrement du commentaire effectué.');
+			$this->redirect("show-post", ["id" => $_POST['post_id']]);
+		} else {
+			$this->redirect("main-home");
+		}
 	}
 	
 	public function delete()
 	{
-		$this->commentManager->delete((int)$_POST['id']);
-		Alert::addAlert(Alert::GREEN, 'Suppression du commentaire effectuée.');
-		header('location: /articles/'.(int)$_POST['post_id']);
+		if($this->isAdmin()){
+			$this->commentManager->delete((int)$_POST['id']);
+			Alert::addAlert(Alert::GREEN, 'Suppression du commentaire effectuée.');
+			$this->redirect("show-post", ["id" => $_POST['post_id']]);
+		} else {
+			$this->redirect("main-home");
+		}
 	}
 	
 	public function hide()
 	{
-		$this->commentManager->hide((int)$_POST['id']);
-		header('location: /articles/'.(int)$_POST['post_id']);
+		if($this->isAdmin()){
+			$this->commentManager->hide((int)$_POST['id']);
+			$this->redirect("show-post", ["id" => $_POST['post_id']]);
+		} else {
+			$this->redirect("main-home");
+		}
 	}
 	
 	public function show()
 	{
-		$this->commentManager->show((int)$_POST['id']);
-		header('location: /articles/'.(int)$_POST['post_id']);
+		if($this->isAdmin()){
+			$this->commentManager->show((int)$_POST['id']);
+			$this->redirect("show-post", ["id" => $_POST['post_id']]);
+		} else {
+			$this->redirect("main-home");
+		}
 	}
 	
 	public function updateForm()
 	{
-		$post_id = (int)$_POST['post_id'];
-		$commentToUpdateId = (int)$_POST['id'];
-		$post = $this->postManager->find($post_id);
-		$comments = $this->commentManager->findByPostId($post_id);
-		$dataPage = [
-			"pageDescription" => "Page d'affichage de l'Article n°" . $post_id,
-			"pageTitle" => "Article " . $post_id,
-			"post" => $post,
-			"comments" => $comments,
-			"commentToUpdateId" => $commentToUpdateId,
-			"view" => PATH_VIEW . "/post.view.php",
-			"template" => PATH_VIEW . "/common/template.php"
-		];
-		$this->generatePage($dataPage);
+		if($this->isAdmin()){
+			$post_id = (int)$_POST['post_id'];
+			$commentToUpdateId = (int)$_POST['id'];
+			$post = $this->postManager->find($post_id);
+			$comments = $this->commentManager->findByPostId($post_id);
+			$dataPage = [
+				"pageDescription" => "Page d'affichage de l'Article n°" . $post_id,
+				"pageTitle" => "Article " . $post_id,
+				"post" => $post,
+				"comments" => $comments,
+				"commentToUpdateId" => $commentToUpdateId,
+				"admin" => true,
+				"view" => PATH_VIEW . "/post.view.php",
+				"template" => PATH_VIEW . "/common/template.php"
+			];
+			$this->generatePage($dataPage);
+		} else {
+			$this->redirect("main-home");
+		}
 	}
 
 	public function update()
 	{
-		$this->commentManager->update((string)$_POST['content'], (int)$_POST['id']);
- 		Alert::addAlert(Alert::GREEN, 'Modification du commentaire effectuée.');
-		$this->redirect("show-post", ["id" => (int)$_POST['post_id']]);
+		if($this->isAdmin()){
+			$this->commentManager->update((string)$_POST['content'], (int)$_POST['id']);
+			Alert::addAlert(Alert::GREEN, 'Modification du commentaire effectuée.');
+		    $this->redirect("show-post", ["id" => (int)$_POST['post_id']]);
+		} else {
+			$this->redirect("main-home");
+		}
 	}
 
 }
